Refactored into own files
