+++ inventree_bulk_plugin/frontend/src/components/BulkDefinitionChildSchemaBuilder.css
@@ -1,0 +1,22 @@
+.collapsable-heading {
+    position: relative;
+    margin-left: 12px;
+}
+  
+.collapsable-heading:before {
+    position: absolute;
+    left: -11px;
+    top: 6px;
+    content: "";
+    display: block;
+    width: 6px;
+    height: 6px;
+    border-top: 2px solid #68686a;
+    border-left: 2px solid #68686a;
+    transform: rotate(135deg);
+    transition: all 0.5s;
+}
+
+.collapsable-heading.active:before {
+      transform: rotate(225deg);
+}
