+++ inventree_bulk_plugin/frontend/src/components/BulkDefinitionSchemaBuilder.tsx
@@ -1,13 +1,9 @@
 import { useState, useCallback, useEffect, useMemo, useId, useRef, StateUpdater } from "preact/hooks"
 import { BulkDefinitionChild, BulkDefinitionChildTemplate, BulkDefinitionSchema, GenerateKeys } from "../utils/types";
-import { ComponentChildren, JSX } from "preact";
+import { JSX } from "preact";
 import { defaultSchema } from "../utils/constants";
-
-const typeHelpTexts = {
-  "boolean": "This must evaluate to something that can be casted to a boolean (e.g. 'true' or 'false').",
-  "number": "This must evaluate to something that can be casted as number.",
-  "text": "",
-}
+import { Input } from "./Input";
+import { BulkDefinitionChildSchemaBuilder } from "./BulkDefinitionChildSchemaBuilder";
 
 interface BulkDefinitionSchemaBuildeProps {
   schema: BulkDefinitionSchema;
@@ -138,312 +134,3 @@ export function BulkDefinitionSchemaBuilder({ schema, setSchema, generateKeys = 
     </div>
   </div>
 }
-
-interface BulkDefinitionChildSchemaBuilderProps {
-  childSchema: BulkDefinitionChild;
-  setChildSchema: StateUpdater<BulkDefinitionChild>;
-  generateKeys: GenerateKeys;
-  extendsKeys: Record<string, string>;
-}
-
-function BulkDefinitionChildSchemaBuilder({ childSchema, setChildSchema, generateKeys, extendsKeys }: BulkDefinitionChildSchemaBuilderProps) {
-  // initially populate childSchema with values
-  useEffect(() => {
-    if (Object.keys(childSchema.generate).length === 0) {
-      setChildSchema(s => ({
-        ...s,
-        generate: Object.fromEntries(Object.entries(generateKeys).filter(([_k, { required }]) => !!required).map(([k, _v]) => [k, ""])),
-      }))
-    }
-  }, [childSchema, generateKeys]);
-
-  // setValue callback for inputs, accepts key and event
-  const setValue = useCallback((key: string) => (e: JSX.TargetedEvent<HTMLInputElement, Event>) => setChildSchema(s => ({ ...s, [key]: e.currentTarget.value })), [setChildSchema]);
-
-  // set a generated value by key and event
-  const setGenerateValue = useCallback((key: string) => (value: string) => setChildSchema(s => ({ ...s, generate: { ...s.generate, [key]: value } })), [setChildSchema]);
-
-  const remainingGenerateKeys = useMemo(() => {
-    const currentKeys = Object.keys(childSchema?.generate || {});
-    return Object.entries(generateKeys).filter(([k, _v]) => !currentKeys.includes(k));
-  }, [generateKeys, childSchema?.generate]);
-
-  const [remainingGenerateKeysValue, setRemainingGenerateKeysValue] = useState("");
-
-  const handleAddGenerateKey = useCallback((key: string) => {
-    if (key !== "") {
-      setGenerateValue(key)("");
-      setRemainingGenerateKeysValue("");
-    }
-  }, [setGenerateValue, setRemainingGenerateKeysValue]);
-
-  const handleDeleteGenerateKey = useCallback((key: string) => () => {
-    setChildSchema(s => ({ ...s, generate: Object.fromEntries(Object.entries(s.generate).filter(([k, _v]) => k !== key)) }))
-  }, []);
-
-  // dimensions
-  const setDimension: (key: "dimensions" | "count") => (i: number) => StateUpdater<string | null> = useCallback((key) => (i) => (newValue) => setChildSchema(s => {
-    s[key][i] = typeof newValue === "function" ? newValue(s[key][i]) : newValue;
-    return { ...s, [key]: s[key] };
-  }), [setChildSchema]);
-  const addDimension = useCallback(() =>
-    setChildSchema(s => {
-      s.dimensions.push("");
-      s.count.push("");
-      return { ...s, dimensions: s.dimensions, count: s.count };
-    }), [setChildSchema]);
-  const removeDimension = useCallback((n: number) => () =>
-    setChildSchema(s => ({ ...s, dimensions: s.dimensions.filter((_, i) => i !== n), count: s.count.filter((_, i) => i !== n) })
-    ), [setChildSchema]);
-
-  // child's childs schemas
-  const setChildChildSchema: (i: number) => StateUpdater<BulkDefinitionChild> = useCallback((i) => (newValue) => setChildSchema(s => {
-    if (!s.childs) s.childs = []
-    s.childs[i] = typeof newValue === "function" ? newValue(s.childs[i]) : newValue;
-    return { ...s, childs: s.childs };
-  }), [setChildSchema]);
-  const addChild = useCallback(() =>
-    setChildSchema(s => {
-      if (!s.childs) s.childs = []
-      s.childs.push(structuredClone(defaultSchema.output));
-      return { ...s, childs: s.childs };
-    }), [setChildSchema]);
-  const removeChild = useCallback((n: number) => () =>
-    setChildSchema(s => {
-      if (!s.childs) return s
-      return ({ ...s, childs: s.childs.filter((_, i) => i !== n) })
-    }
-    ), [setChildSchema]);
-
-  // collapsable handling for advanced child options
-  const id = useId();
-  const outputAdvancedId = `bulk-child-schema-editor-${id}`;
-  const [outputAdvancedState, setOutputAdvancedState] = useState(false);
-
-  const onOutputAdvanceToggle = useCallback(() => {
-    setOutputAdvancedState(oldState => {
-      // @ts-ignore: correct types for bootstrap table are not available
-      $(`#${outputAdvancedId}`).collapse(!oldState ? "show" : "hide");
-      return !oldState;
-    });
-  }, []);
-
-  if (childSchema === null) {
-    return <></>;
-  }
-
-  return (
-    <div class="">
-      <Input label="Dimensions" tooltip="A childs naming convention could have multiple dimensions where it is counting in" type="array" onDelete={removeDimension} value={childSchema.dimensions.map(x => x || "")} onInput={setDimension("dimensions")} onAdd={addDimension} />
-      <Input label="Count" tooltip="Limit the amount of generated items for the individual dimension" type="array" value={childSchema.count.map(x => x || "")} onInput={setDimension("count")} />
-
-      <h6 class={`user-select-none collapsable-heading ${outputAdvancedState ? "active" : ""}`} role="button" onClick={onOutputAdvanceToggle}>Advanced</h6>
-      <div class="collapse" id={outputAdvancedId}>
-        <Input label="Parent name match" tooltip="First child that matches the parent name matcher will be chosen for generating the childs for a specific parent. Must evalueate to something that can be casted to a boolean." type="text" value={childSchema.parent_name_match || ""} onInput={setValue("parent_name_match")} />
-        {extendsKeys && <Input label="Extends" tooltip="Choose to extend from a template" type="select" options={extendsKeys} value={childSchema.extends || ""} onInput={setValue("extends")} />}
-      </div>
-
-      <div class="mt-3">
-        <Tooltip text={`You can use {{dim.x}} as a placeholder for the generated output of the dimension. For more info see the Readme.`}>
-          <h5>Generate</h5>
-        </Tooltip>
-      </div>
-      {Object.entries(generateKeys).filter(([key, _definition]) => childSchema.generate[key] !== undefined).map(([key, { name, required, type }]) => (
-        <Input label={name} tooltip={typeHelpTexts[type]} type="text" value={childSchema.generate[key]} onInput={(e: JSX.TargetedEvent<HTMLInputElement, Event>) => setGenerateValue(key)(e.currentTarget.value)} onDelete={required ? undefined : handleDeleteGenerateKey(key)} />
-      ))}
-      {remainingGenerateKeys.length > 0 && (
-        <div style="display: flex; max-width: 200px;">
-          <select class="form-select form-select-sm" value={remainingGenerateKeysValue} onInput={(e) => handleAddGenerateKey(e.currentTarget.value)}>
-            <option selected value="">Add key</option>
-            {remainingGenerateKeys.map(([key, { name }]) => <option value={key}>{name}</option>)}
-          </select>
-        </div>
-      )}
-
-      <h5 class="mt-3">Childs</h5>
-      <div class="ms-3">
-        {childSchema.childs?.map((child, i) => (
-          <div class="card mb-2">
-            <div class="d-flex justify-content-between">
-              <div class="col p-3">
-                <BulkDefinitionChildSchemaBuilder childSchema={child} setChildSchema={setChildChildSchema(i)} generateKeys={generateKeys} extendsKeys={extendsKeys} />
-              </div>
-              <div class="p-1">
-                <button onClick={removeChild(i)} class="btn btn-outline-danger">X</button>
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-      <button onClick={addChild} class="btn btn-outline-primary btn-sm">Add child</button>
-    </div>
-  )
-}
-
-
-interface DefaultInputProps {
-  label: string;
-  tooltip?: string;
-  [key: string]: any;
-}
-
-interface TextInputProps extends DefaultInputProps {
-  type: "text" | "email" | "number";
-  value: string;
-  onDelete?: () => void;
-}
-
-interface NumberInputProps extends DefaultInputProps {
-  type: "number";
-  value: number;
-  onDelete?: () => void;
-}
-
-interface CheckboxInputProps extends DefaultInputProps {
-  type: "checkbox";
-  value: boolean;
-}
-
-interface SelectInputProps extends DefaultInputProps {
-  type: "select";
-  value: string;
-  options: Record<string, string>
-}
-
-interface ArrayInputProps extends DefaultInputProps {
-  type: "array";
-  value: string[];
-  onDelete?: (i: number) => () => void;
-  onAdd?: () => void
-}
-
-type InputProps = TextInputProps | NumberInputProps | CheckboxInputProps | SelectInputProps | ArrayInputProps;
-
-export function Input(props: InputProps) {
-  const id = useId();
-
-  return <div class="form-group row">
-    <div class="col-sm-2">
-      <Tooltip text={props.tooltip}>
-        <label for={`input-${id}`} class="col-form-label col-form-label-sm">{props.label}</label>
-      </Tooltip>
-    </div>
-    {(() => {
-      if (props.type === "text" || props.type === "number" || props.type === "email") {
-        const { type, ...inputProps } = props;
-        return (
-          <div class="col-sm-10 input-group" style="flex: 1;">
-            <input type={type} class="form-control form-control-sm" id={`input-${id}`} {...inputProps} />
-            {props.onDelete && <button class="btn btn-outline-danger btn-sm" onClick={props.onDelete}>X</button>}
-          </div>
-        )
-      } else if (props.type === "checkbox") {
-        const { type, value, ...inputProps } = props;
-        return (
-          <div class="col-sm-10">
-            <input type={type} class="form-check-input form-check-sm" id={`input-${id}`} checked={value} {...inputProps} />
-          </div>
-        )
-      } else if (props.type === "select") {
-        return (
-          <div class="col-sm-10">
-            <select class="form-select form-select-sm" id={`input-${id}`} {...props}>
-              {Object.entries(props.options).map(([k, v]) => <option value={k}>{v}</option>)}
-            </select>
-          </div>
-        )
-      } else if (props.type === "array") {
-        const { onDelete, onAdd } = props;
-        return (
-          <div class="col-sm-10">
-            <div class="d-flex flex-row flex-nowrap" style="gap: 4px">
-              {props.value.map((v: string, i: number) => <div class="input-group" style={`max-width: 350px; ${!props.onAdd && props.value.length === i + 1 ? "margin-right: 31px;" : ""}`}>
-                <input type="text" class="form-control form-control-sm" id={`input-${id}-${i}`} value={v} onInput={(e) => props.onInput(i)(() => e.currentTarget.value)} />
-                {onDelete && <button class="btn btn-outline-danger btn-sm" onClick={onDelete(i)}>X</button>}
-              </div>
-              )}
-              {onAdd && <button class="btn btn-outline-primary btn-sm" onClick={() => onAdd()}>+</button>}
-            </div>
-          </div>
-        )
-      } else {
-        return "Not implemented";
-      }
-    })()}
-  </div>
-}
-
-interface TooltipProps {
-  text?: string;
-  children: ComponentChildren;
-}
-export function Tooltip({ text, children }: TooltipProps) {
-  if (!text) {
-    return <>{children}</>;
-  }
-
-  return <div class="preact-tooltip">
-    {children}
-    <span class="preact-tooltiptext">{text}</span>
-  </div>
-
-}
-
-const styles = `
-/* Tooltips - (Prefixed so the classes doesn't interfer with bootstraps classes) */
-.preact-tooltip {
-  position: relative;
-  display: inline-block;
-}
-
-.preact-tooltip .preact-tooltiptext {
-  visibility: hidden;
-  width: 350px;
-  background-color: black;
-  color: #fff;
-  border-radius: 6px;
-  padding: 5px;
-  opacity: 0;
-  transition: 50ms all;
-  
-  /* Position the tooltip */
-  position: absolute;
-  z-index: 1000;
-  top: -5px;
-  left: 105%;
-}
-
-.preact-tooltip:hover .preact-tooltiptext {
-  visibility: visible;
-  opacity: 1;
-}
-
-.collapsable-heading {
-  position: relative;
-  margin-left: 12px;
-}
-
-.collapsable-heading:before {
-  position: absolute;
-  left: -11px;
-  top: 6px;
-  content: "";
-  display: block;
-  width: 6px;
-  height: 6px;
-  border-top: 2px solid #68686a;
-  border-left: 2px solid #68686a;
-  transform: rotate(135deg);
-  transition: all 0.5s;
-}
-.collapsable-heading.active:before {
-	transform: rotate(225deg);
-}
-
-.loadingindicator {
-  display: none;
-}
-`;
-const styleSheet = document.createElement("style");
-styleSheet.innerHTML = styles;
-document.head.appendChild(styleSheet);
