+++ inventree_bulk_plugin/frontend/src/components/Tooltip.css
@@ -1,0 +1,27 @@
+/* Tooltips - (Prefixed so the classes doesn't interfer with bootstraps classes) */
+.preact-tooltip {
+    position: relative;
+    display: inline-block;
+}
+  
+.preact-tooltip .preact-tooltiptext {
+    visibility: hidden;
+    width: 350px;
+    background-color: black;
+    color: #fff;
+    border-radius: 6px;
+    padding: 5px;
+    opacity: 0;
+    transition: 50ms all;
+
+    /* Position the tooltip */
+    position: absolute;
+    z-index: 1000;
+    top: -5px;
+    left: 105%;
+}
+
+.preact-tooltip:hover .preact-tooltiptext {
+    visibility: visible;
+    opacity: 1;
+}
